Finished basic Back-end

+++ L1Q1 Project Smart Energy/SmartEnergy.Library/Measurements/Models/Laadsessie.cs
@@ -2,6 +2,6 @@ namespace SmartEnergy.Library.Measurements.Models;
 public record Laadsessie(DateTime Start,
                           DateTime Stop,
                           TimeSpan Duur,
-                          decimal Prijs,
+                          double Prijs,
                           double TotaalKWh
                           );
